requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.0
Pillow==9.2.0
requests==2.27.1
